Refactor menu system and consolidate SWOT/Gantt into project module
- Add dedicated "Aktivní" section at top of sidebar for active module
- Move SWOT templates and views into project module
- Add Gantt diagram view and template to project module
- Remove redundant JavaScript for moving active module
- Add CSS styling for active section in sidebar
- Update URLs and imports for consolidated structure
