+++ fdk_cz/templates/inc/menus/project_menu.html
@@ -14,6 +14,9 @@
     <a href="{% url 'list_swot_analyses' %}" class="nav-submenu-item">
       <span class="nav-icon">&#128202;</span> SWOT analýzy
     </a>
+    <a href="{% url 'gantt_chart' %}" class="nav-submenu-item">
+      <span class="nav-icon">&#128197;</span> Ganttův diagram
+    </a>
     <a href="{% url 'all_project_logs' %}" class="nav-submenu-item">
       <span class="nav-icon">&#128196;</span> Log aktivit
     </a>
