+++ fdk_cz/templates/project/swot_create.html
