+++ fdk_cz/templates/project/swot_edit.html
