+++ fdk_cz/templates/project/swot_list.html
